Adds configurable base domain for tunnel URLs
Allows users to specify a base domain for generating tunnel URLs.

This change introduces a new `baseDomain` parameter to the client, allowing customization of the generated tunnel URL.
It defaults to `jiboia.cloud` or the value of the `TUNNEL_DOMAIN` environment variable, if set.

This provides more flexibility and control over the tunnel endpoint.
Also fixes a typo error in client log message.